feat: enforce stricter casting rules across crates
- Deny various Clippy lint warnings related to potential truncation, sign loss, and wrapping casts at the crate boundaries in `ultrasql-core`, `ultrasql-executor`, `ultrasql-mvcc`, `ultrasql-optimizer`, `ultrasql-parser`, `ultrasql-planner`, `ultrasql-protocol`, `ultrasql-server`, `ultrasql-storage`, `ultrasql-txn`, `ultrasql-vec`, and `ultrasql-wal`.
- Introduce `#[allow(...)]` annotations in test modules and specific functions where necessary to accommodate known safe casts, particularly in index arithmetic and fixed-width packing scenarios.
- Update casting logic to use `try_from` where applicable to ensure safe conversions and prevent panics on invalid casts.
- Ensure that all changes maintain the integrity of the on-disk format and prevent silent truncation in critical areas of the codebase.

+++ crates/ultrasql-optimizer/src/enumeration/physical_selection.rs
@@ -175,8 +175,8 @@ pub fn select_scan_physical(
     stats: &dyn StatsSource,
     _gucs: &CostGucs,
 ) -> PhysicalOp {
-    let total_rows = stats.row_count(table) as f64;
-    if total_rows == 0.0 {
+    let total_rows = stats.row_count(table);
+    if total_rows == 0 {
         return PhysicalOp::SeqScan;
     }
 
@@ -190,7 +190,7 @@ pub fn select_scan_physical(
         for pred in predicates {
             if predicate_references_column(pred, leading_col) {
                 let sel =
-                    crate::cost::selectivity::selectivity(pred, stats, table, total_rows as u64);
+                    crate::cost::selectivity::selectivity(pred, stats, table, total_rows);
                 matches.push((hint, sel));
                 break; // one predicate per index is enough
             }
